Module Dolmen_loop.Alarm

Some types

exception Out_of_time

exception Out_of_space

module type S = sig ... end

The module signature that implements everything needed to use alarms for time and sizes.

type t = (module S)

An alarm system is a packed module.

Alarm Implementations

val dummy : t

Dummy implementation. Does nothing.

val linux : t

Linux alarms. Uses Gc alarm for size limits, and Unix timers for time limits. Fairly accurate and reliable.

val windows : t

Windows alarms. Uses Gc alarm for size limits as well as time limits. The time limit is not enforced very precisely/reliably because of the use of Gc alarms.

val default : t

A default alarm, chosen based on `Sys.os_type`. Can be either

nux

}

or windows, and falls back on dummy for unknown os (e.g. js_of_ocaml).