Module type Alarm.S

The module signature that implements everything needed to use alarms for time and sizes.

type t

The type of active alarms.

val setup : time:float -> size:float -> t

Create an alarm ith the given size and time limits.

val delete : t -> unit

Delete an alarm.