Module Dolmen_loop.Pipeline

Pipelines

This module implements functorised pipelines. A pipeline is basically a series of computations that is meant to be executed on a stream of elements. Each computation in a pipeline is called a pipe. A pipe can be as simple as a mapping function, transforming the input element into an output element, or more complex, expanding an element into a sequence of elemnents, and performing a fixpoint expansion.

exception Sigint

Alias to Sys.Break. Raised upon user interrupt / when a Sys.sigint signal is received.

exception Out_of_time

Alias to Alarm.Out_of_time.

exception Out_of_space

Alias to Alarm.Out_of_space.

module Make (State : State.S) : sig ... end

Concrete pipelines.