Module S.Parser

The Parser module for the language.

Main interface

exception Error of int

Exception raised by the parser when it encounters an error.

val file : (Lexing.lexbuf -> token) -> Lexing.lexbuf -> Dolmen.Std.Answer.t list

A function that parses an entire file, i.e until the end-of-file token, and return the list of parsed statements.

val input : (Lexing.lexbuf -> token) -> Lexing.lexbuf -> Dolmen.Std.Answer.t option

A function to parse a single statement. Returns None if it encounters the end-of-file token. Used for incremental parsing.