Module type State_intf.Parser_pipe

This modules defines the smallest signatures for a solver state that allow to instantiate the Parser.Pipe functor.

type t

The type of state

type term

The type of solver terms.

val warn : ?loc:Dolmen.Std.Loc.full -> t -> ('a, Stdlib.Format.formatter, unit, t) Stdlib.format4 -> 'a

Emit a warning

val input_file_loc : t -> Dolmen.Std.Loc.file

CUrrent input file location meta-data.

val set_input_file_loc : t -> Dolmen.Std.Loc.file -> t

Set the input file location meta-data.

val start : phase -> unit

Hook at the start of a phase

val stop : phase -> unit

Hook at the end of a phase

val prelude : t -> string

Some prelude to print at the begining of lines when in interactive mode.

val is_interactive : t -> bool

Whether we are running in interactive mode.

val set_mode : t -> mode -> t

val set_lang : t -> Logic.language -> t

Set the input language.

val input_mode : t -> mode option

Return the current mode (if any).

val input_lang : t -> Logic.language option

Return the input language (if any).

val input_dir : t -> string

Return the directory of the input source (e.g. the directory of the input file, or the current directory if in interactive mode).

val input_source : t -> source

Return the input source.

val file_not_found : loc:Dolmen.Std.Loc.full -> dir:string -> file:string -> 'a

Callback for when a file specified by the input source is not found.