Module Typer.Ext

Define typing extensions.

These extensions are typically extensions used by some community, but not yet part of the standard.

type t

The type of typing extensions.

val name : t -> string

Extension name, sould be suitable for cli options.

val builtins : t -> Typer_intf.lang -> T.builtin_symbols

Returns the typing builtins from an extension.

val create : name:string -> builtins:(Typer_intf.lang -> T.builtin_symbols) -> t

Create a new extension.

val iter : (t -> unit) -> unit

Iterate over all the registered extensions.

val find_all : string -> t list

Returns the extensions that have been registered with the given name.

val bvconv : t

Typing extension to add `bv2nat` and `int2bv`.