Module Typer.Make

Parameters

module Expr : Expr_intf.S

module Print : Expr_intf.Print with type ty := Expr.ty and type ty_var := Expr.ty_var and type ty_cst := Expr.ty_cst and type ty_def := Expr.ty_def and type term := Expr.term and type term_var := Expr.term_var and type term_cst := Expr.term_cst and type formula := Expr.formula

module State : State.S

module Typer : Typer with type state := State.t and type ty := Expr.ty and type ty_var := Expr.ty_var and type ty_cst := Expr.ty_cst and type ty_def := Expr.ty_def and type term := Expr.term and type term_var := Expr.term_var and type term_cst := Expr.term_cst and type formula := Expr.formula

Signature

include Typer_intf.Types with type state := State.t with type env = Typer.env with type ty := Expr.ty with type ty_var := Expr.ty_var with type ty_cst := Expr.ty_cst with type ty_def := Expr.ty_def with type term := Expr.term with type term_var := Expr.term_var with type term_cst := Expr.term_cst with type formula := Expr.formula

type env = Typer.env

val type_check : bool State.key

val init : type_check:bool -> State.t -> State.t

Types

type +'a stmt = {

id : Dolmen.Std.Id.t;

loc : Dolmen.Std.Loc.t;

contents : 'a;

attrs : Dolmen.Std.Term.t list;

implicit : bool;

}

Wrapper around statements. It records implicit type declarations.

type decl = [

| `Type_decl of Expr.ty_cst * Expr.ty_def option

| `Term_decl of Expr.term_cst

]

The type of top-level type declarations.

type decls = [

| `Decls of decl list

]

A list of type declarations.

type def = [

| `Type_alias of Dolmen.Std.Id.t * Expr.ty_cst * Expr.ty_var list * Expr.ty

| `Term_def of Dolmen.Std.Id.t * Expr.term_cst * Expr.ty_var list * Expr.term_var list * Expr.term

| `Instanceof of Dolmen.Std.Id.t * Expr.term_cst * Expr.ty list * Expr.ty_var list * Expr.term_var list * Expr.term

]

The type of top-level type definitions. Type definitions are inlined and so can be ignored.

type defs = [

| `Defs of def list

]

A list of definitions

type assume = [

| `Hyp of Expr.formula

| `Goal of Expr.formula

| `Clause of Expr.formula list

]

The type of top-level assertion statements

type solve = [

| `Solve of Expr.formula list * Expr.formula list
(*
`Solve (hyps, goals) represents a sequent with local hypotheses hyps and local goals goals.
*)

]

Top-level solve instruction

type get_info = [

| `Get_info of string

| `Get_option of string

| `Get_proof

| `Get_unsat_core

| `Get_unsat_assumptions

| `Get_model

| `Get_value of Expr.term list

| `Get_assignment

| `Get_assertions

| `Echo of string

| `Other of Dolmen.Std.Id.t * Dolmen.Std.Statement.term list

]

Various info getters

type set_info = [

| `Set_logic of string * Dolmen_type.Logic.t

| `Set_info of Dolmen.Std.Statement.term

| `Set_option of Dolmen.Std.Statement.term

]

Info setters

type stack_control = [

| `Pop of int

| `Push of int

| `Reset_assertions

| `Reset

]

Stack control

type exit = [

| `Exit

]

Exit statement

type typechecked = [

| defs

| decls

| assume

| solve

| get_info

| set_info

| stack_control

| exit

]

The type of statements after typechecking

val print : Format.formatter -> typechecked stmt -> unit

Printing funciton for typechecked statements.

val check : State.t -> Dolmen.Std.Statement.t -> State.t * typechecked stmt list

Typechecks a statement. This returns a list because some languages have implicit declarations and definitions: for intance, tptp has symbol inferrence (which result in implicit declarations), and smtlib has `:named` annotation (which result in implicit definitions). The list is returned in dependency order: implicits declarations/definitions are first in the list, and they should be in an order such that only later statements can depend on earlier statements.

val typecheck : State.t -> Dolmen.Std.Statement.t -> State.t * [ `Continue of typechecked stmt list | `Done of unit ]

Typechecks a statement.