Module T.Const

A module for constant symbols that occur in terms.

type t = Dolmen_std.Expr.term_cst

The type of constant symbols that can occur in terms

val compare : t -> t -> int

Comparison function on constant symbols.

val arity : t -> int * int

Returns the arity of a term constant.

val mk : string -> ty -> t

Create a constant symbol.

val set_tag : t -> 'a tag -> 'a -> unit

Tag a constant.