Module type Typer_intf.Pipe_res

This modules defines the result signature of the Typer.Pipe functor

include Pipe_types

type state

type ty

type ty_var

type ty_cst

type term

type term_var

type term_cst

type formula

Types

type +'a stmt = {

id : Dolmen.Std.Id.t;

loc : Dolmen.Std.Loc.t;

contents : 'a;

}

Wrapper around statements. It records implicit type declarations.

type decl = [

| `Type_decl of ty_cst

| `Term_decl of term_cst

]

The type of top-level type declarations.

type decls = [

| `Decls of decl list

]

A list of type declarations.

type def = [

| `Type_def of Dolmen.Std.Id.t * ty_cst * ty_var list * ty

| `Term_def of Dolmen.Std.Id.t * term_cst * ty_var list * term_var list * term

]

The type of top-level type definitions. Type definitions are inlined and so can be ignored.

type defs = [

| `Defs of def list

]

A list of definitions

type assume = [

| `Hyp of formula

| `Goal of formula

| `Clause of formula list

]

The type of top-level assertion statements

type solve = [

| `Solve of formula list

]

Top-level solve instruction

type get_info = [

| `Get_info of string

| `Get_option of string

| `Get_proof

| `Get_unsat_core

| `Get_unsat_assumptions

| `Get_model

| `Get_value of term list

| `Get_assignment

| `Get_assertions

| `Echo of string

| `Plain of Dolmen.Std.Statement.term

]

Various info getters

type set_info = [

| `Set_logic of string

| `Set_info of Dolmen.Std.Statement.term

| `Set_option of Dolmen.Std.Statement.term

]

Info setters

type stack_control = [

| `Pop of int

| `Push of int

| `Reset_assertions

| `Reset

| `Exit

]

Stack control

type typechecked = [

| defs

| decls

| assume

| solve

| get_info

| set_info

| stack_control

]

The type of statements after typechecking

val print : Stdlib.Format.formatter -> typechecked stmt -> unit

Printing funciton for typechecked statements.

val typecheck : state -> Dolmen.Std.Statement.t -> state * [ `Continue of typechecked stmt | `Done of unit ]

Typechecks a statement.