Parameter Pipe.2-State

type t

The type of state

type term

The type of solver terms.

val warn : ?⁠loc:Dolmen_std.Loc.full -> t -> ('a, Stdlib.Format.formatter, unit, t) Stdlib.format4 -> 'a

Emit a warning

val input_file_loc : t -> Dolmen_std.Loc.file

CUrrent input file location meta-data.

val set_input_file_loc : t -> Dolmen_std.Loc.file -> t

Set the input file location meta-data.

val start : Dolmen_loop.State_intf.phase -> unit

Hook at the start of a phase

val stop : Dolmen_loop.State_intf.phase -> unit

Hook at the end of a phase

val prelude : t -> string

Some prelude to print at the begining of lines when in interactive mode.

val is_interactive : t -> bool

Whether we are running in interactive mode.

val set_mode : t -> Dolmen_loop.State_intf.mode -> t

val set_lang : t -> Dolmen_loop.Logic.language -> t

Set the input language.

val input_mode : t -> Dolmen_loop.State_intf.mode option

Return the current mode (if any).

val input_lang : t -> Dolmen_loop.Logic.language option

Return the input language (if any).

val input_dir : t -> string

Return the directory of the input source (e.g. the directory of the input file, or the current directory if in interactive mode).

val input_source : t -> Dolmen_loop.State_intf.source

Return the input source.

val file_not_found : loc:Dolmen_std.Loc.full -> dir:string -> file:string -> 'a

Callback for when a file specified by the input source is not found.