Module Dolmen_loop__State_intf

Solver State

This module defines various interfaces for states used in main loops of solvers (or other binaries) using dolmen.

Useful type aliases

type source = [

| `Stdin

| `File of string

| `Raw of string * string

]

type phase = [

| `Parsing

| `Include

| `Typing

| `Solving

]

type mode = [

| `Full

| `Incremental

]

Signatures

module type Pipeline = sig ... end

This modules defines the smallest signatures for a solver state that allow to instantiate the Pipeline.Make functor.

module type Parser_pipe = sig ... end

This modules defines the smallest signatures for a solver state that allow to instantiate the Parser.Pipe functor.

module type Typer = sig ... end

This modules defines the smallest signatures for a solver state that allow to instantiate the Typer.Make functor.

module type Typer_pipe = sig ... end

This modules defines the smallest signatures for a solver state that allow to instantiate the Typer.Pipe functor.

module type Header_pipe = sig ... end

This modules defines the smallest signatures for a solver state that allow to instantiate the Headers.Pipe functor.