Module type Dolmen_loop__State_intf.Pipeline

This modules defines the smallest signatures for a solver state that allow to instantiate the Pipeline.Make functor.

type t

The type of values recording options for the current run.

val time_limit : t -> float

The time limit for one original statement (in seconds).

val size_limit : t -> float

The size limit for one original statement (in octets).