Module T.Ty

type t = Dolmen_std.Expr.ty

The type of types.

type 'a tag

A type for tags to attach to arbitrary types.

module Var : sig ... end with type t = Dolmen_std.Expr.ty_var

A module for variables that occur in types.

module Const : sig ... end with type t = Dolmen_std.Expr.ty_const

A module for constant symbols the occur in types.

val prop : t

The type of propositions

val of_var : Var.t -> t

Create a type from a variable.

val apply : Const.t -> t list -> t

Application for types.

val wildcard : unit -> t

Create a fresh type wildcard.

val tag : t -> 'a tag -> 'a -> unit

Annotate the given type with the given tag and value.