Module Dolmen_loop__Typer_intf

Typer

module type Pipe_types = sig ... end

module type Pipe_arg = sig ... end

This modules defines the smallest signatures for a typechecker that allow to instantiate the Typer.Pipe functor.

module type S = sig ... end

This modules defines the result signature of the Typer.Make functor

module type Pipe_res = sig ... end

This modules defines the result signature of the Typer.Pipe functor