Module Dolmen_loop.Headers

Header fields

module Field : sig ... end

Header set

type t

Header set, i.e. a map of fields to values for the header.

val empty : t

The empty header set

val set : t -> Field.t -> string -> t

Add/set a header to the corresponding value.

val get : t -> Field.t -> string option

Get a header value, if present in the set.

val mem : t -> Field.t -> bool

Test the presence of a header field.

val remove : t -> Field.t -> t

Remove a field from a set.

Pipe functor

module Pipe : functor (State : State_intf.Header_pipe with type header_state := t) -> Headers_intf.S with type state := State.t