Module Headers.Make

Parameters

module S : State.S

Signature

val header_check : bool S.key

Key for deciding whether to do any check on headers.

val header_state : t S.key

Key for the local header state.

val header_licenses : string list S.key

Key for the list of allowed licenses in headers.

val header_lang_version : string option S.key

Key for the allowed language version in headers.

val init : header_check:bool -> ?header_state:t -> header_licenses:string list -> header_lang_version:string option -> S.t -> S.t

Init a state with all the relevant keys for this pipeline.

val inspect : S.t -> Dolmen.Std.Statement.t -> S.t * Dolmen.Std.Statement.t

Check the headers

val check : S.t -> S.t

Check a state for the required headers (once a whole pipeline has been completed