Module Dolmen_loop.Pipeline

Pipelines

This module implements functorised pipelines. A pipeline is basically a series of computations that is meant to be executed on a stream of elements. Each computation in a pipeline is called a pipe. A pipe can be as simple as a mapping function, transforming the input element into an output element, or more complex, expanding an element into a sequence of elemnents, and performing a fixpoint expansion.

exception Sigint

exception Out_of_time

exception Out_of_space

module Make (State : State_intf.Pipeline) : sig ... end

Concrete pipelines.