Module Typer.Make

Parameters

S : State_intf.Typer with type ty_state := ty_state

Signature

type state

The type of state for a whole pipeline.

type ty_state

The type for the state of the typer.

type env

The type of the typechecker environment.

type 'a fragment

The type of fragments on which error/warning can occur.

type error

The type of type-checking errors.

type warning

The type of type-checking warnings.

type builtin_symbols

The type of builint symbols for the type-checker.

This signature includes the requirements to instantiate the {Pipes.Make: functor

include Typer_intf.Pipe_arg with type state := state and type ty := Dolmen_std.Expr.ty and type ty_var := Dolmen_std.Expr.ty_var and type ty_const := Dolmen_std.Expr.ty_const and type term := Dolmen_std.Expr.term and type term_var := Dolmen_std.Expr.term_var and type term_const := Dolmen_std.Expr.term_const and type formula := Dolmen_std.Expr.formula

include Typer_intf.Pipe_types

type state

type ty

type ty_var

type ty_const

type term

type term_var

type term_const

type formula

val typecheck : state -> bool

val reset : state -> ?⁠loc:Dolmen_std.Loc.t -> unit -> state

val push : state -> ?⁠loc:Dolmen_std.Loc.t -> int -> state

val pop : state -> ?⁠loc:Dolmen_std.Loc.t -> int -> state

val set_logic : state -> ?⁠loc:Dolmen_std.Loc.t -> string -> state

val defs : state -> ?⁠loc:Dolmen_std.Loc.t -> ?⁠attr:Dolmen_std.Term.t -> Dolmen_std.Statement.defs -> state * [ `Type_def of Dolmen_std.Id.t * ty_var list * ty | `Term_def of Dolmen_std.Id.t * term_const * ty_var list * term_var list * term ] list

val decls : state -> ?⁠loc:Dolmen_std.Loc.t -> ?⁠attr:Dolmen_std.Term.t -> Dolmen_std.Statement.decls -> state * [ `Type_decl of ty_const | `Term_decl of term_const ] list

val terms : state -> ?⁠loc:Dolmen_std.Loc.t -> ?⁠attr:Dolmen_std.Term.t -> Dolmen_std.Term.t list -> state * term list

val formula : state -> ?⁠loc:Dolmen_std.Loc.t -> ?⁠attr:Dolmen_std.Term.t -> goal:bool -> Dolmen_std.Term.t -> state * formula

val formulas : state -> ?⁠loc:Dolmen_std.Loc.t -> ?⁠attr:Dolmen_std.Term.t -> Dolmen_std.Term.t list -> state * formula list

val print_fragment : Stdlib.Format.formatter -> (env * 'a fragment) -> unit

Print a code fragment

val report_error : Stdlib.Format.formatter -> error -> unit

Report a typing error on the given formatter.

val report_warning : warning -> (Stdlib.Format.formatter -> unit -> unit) option

Return a reporter for the given warning, if the warning should be reported.

val additional_builtins : builtin_symbols Stdlib.ref

This reference can be modified to parse new builtin symbols. By default no additional builtin symbols are parsed. It is added for all the languages except Dimacs, and iCNF.